requirements/prod.txt
aiogram==2.9
aiohttp-socks==0.3.9
python-dotenv==0.13.0
